refactor(task-service): integrate centralized ApiService in TaskService
