+++ src/app/features/tasks-dashboard/services/task.service.ts
@@ -1,24 +1,22 @@
 import { Injectable } from '@angular/core';
-import { HttpClient, HttpParams } from '@angular/common/http';
+import { HttpParams } from '@angular/common/http';
 import { Observable, catchError, throwError } from 'rxjs';
 import {
-  ApiResponse,
   GroupedTasksResponse,
   Task,
   TaskPaginationParams,
   SuggestedTasksParams,
 } from '../../../core/models/tasks-model';
-import { environment } from '../../../../environments/environment';
+import { ApiResponse } from '../../../core/models/api.model';
+import { ApiService } from '../../../core/services/api/api-service';
 import { ErrorHandlerService } from '../../../core/services/error/error-handler';
 
 @Injectable({
   providedIn: 'root',
 })
 export class TaskService {
-  private readonly apiUrl = `${environment.url}/api/v1/tasks`;
-
   constructor(
-    private http: HttpClient,
+    private apiService: ApiService,
     private errorHandler: ErrorHandlerService,
   ) {}
 
@@ -31,8 +29,8 @@ export class TaskService {
       .set('completedPage', (params.completedPage ?? 0).toString())
       .set('completedSize', (params.completedSize ?? 10).toString());
 
-    return this.http
-      .get<ApiResponse<GroupedTasksResponse>>(`${this.apiUrl}/my-tasks`, { params: httpParams })
+    return this.apiService
+      .get<ApiResponse<GroupedTasksResponse>>('/tasks/my-tasks', { params: httpParams })
       .pipe(catchError(this.handleError));
   }
 
@@ -42,14 +40,14 @@ export class TaskService {
       .set('taskType', params.taskType)
       .set('limit', (params.limit ?? 5).toString());
 
-    return this.http
-      .get<ApiResponse<Task[]>>(this.apiUrl, { params: httpParams })
+    return this.apiService
+      .get<ApiResponse<Task[]>>('/tasks', { params: httpParams })
       .pipe(catchError(this.handleError));
   }
 
   public getTaskById(id: string): Observable<ApiResponse<Task>> {
-    return this.http
-      .get<ApiResponse<Task>>(`${this.apiUrl}/${id}`)
+    return this.apiService
+      .get<ApiResponse<Task>>(`/tasks/${id}`)
       .pipe(catchError(this.handleError));
   }
